wip: adding modes
