+++ src/report-templates/regression/index.js
@@ -5,6 +5,11 @@ import outdent from 'outdent';
 import {
 	c, sub, sup, strong,
 } from '../../lib/markdown.js';
+import {
+	getSizeLabels,
+	parseDisplaySize,
+	listSizes,
+} from '../utils.js';
 import comparePackages from './compare-packages.js';
 
 const directionSymbol = (value) => {
@@ -20,25 +25,6 @@ const directionSymbol = (value) => {
 };
 
 const formatDelta = ({ delta, percent }) => (delta ? (percent + directionSymbol(delta)) : '');
-
-const supportedSizes = {
-	uncompressed: {
-		label: 'Size',
-		property: 'size',
-	},
-	gzip: {
-		label: 'Gzip',
-		property: 'sizeGzip',
-	},
-	brotli: {
-		label: 'Brotli',
-		property: 'sizeBrotli',
-	},
-};
-
-const listSizes = (displaySizes, callback) => displaySizes
-	.map(({ property }) => callback(property))
-	.join(' / ');
 
 function generateComment({
 	headPkgData,
@@ -58,16 +44,8 @@ function generateComment({
 	setOutput('pkgComparisonData', pkgComparisonData);
 
 	const { changed, unchanged, hidden } = pkgComparisonData.files;
-	const displaySizes = displaySize
-		.split(',')
-		.map(s => s.trim())
-		.filter(s => supportedSizes.hasOwnProperty(s)) // eslint-disable-line no-prototype-builtins
-		.map(s => supportedSizes[s]);
-
-	let sizeHeadingLabel = '';
-	if (displaySizes.length > 1 || displaySizes[0].property !== 'size') {
-		sizeHeadingLabel = ` (${displaySizes.map(s => s.label).join(' / ')})`;
-	}
+	const displaySizes = parseDisplaySize(displaySize);
+	const sizeHeadingLabel = getSizeLabels(displaySizes);
 
 	const table = markdownTable([
 		['File', `Before${sizeHeadingLabel}`, `After${sizeHeadingLabel}`],
