+++ src/report-templates/utils.js
@@ -9,6 +9,43 @@ function partionHidden(hideFilesGlob, files) {
 	return partition(files, file => hideFilesPtrn.test(file.path));
 }
 
+function getSizeLabels(displaySizes) {
+	if (displaySizes.length === 1 && displaySizes[0].property === 'size') {
+		return '';
+	}
+	return ` (${displaySizes.map(s => s.label).join(' / ')})`;
+}
+
+const supportedSizes = {
+	uncompressed: {
+		label: 'Size',
+		property: 'size',
+	},
+	gzip: {
+		label: 'Gzip',
+		property: 'sizeGzip',
+	},
+	brotli: {
+		label: 'Brotli',
+		property: 'sizeBrotli',
+	},
+};
+
+function parseDisplaySize(displaySize) {
+	return displaySize
+		.split(',')
+		.map(s => s.trim())
+		.filter(s => supportedSizes.hasOwnProperty(s)) // eslint-disable-line no-prototype-builtins
+		.map(s => supportedSizes[s]);
+}
+
+const listSizes = (displaySizes, callback) => displaySizes
+	.map(({ property }) => callback(property))
+	.join(' / ');
+
 export {
-	partionHidden, // eslint-disable-line import/prefer-default-export
+	partionHidden,
+	getSizeLabels,
+	parseDisplaySize,
+	listSizes,
 };
